Broadcast fixes, Utils, env setup
