+++ load-test/send_load.sh
@@ -1,3 +1,22 @@
 #/bin/sh
 
-python3 ../python-qa-tools/load-test/send_load.py $1 $2
+set -e
+
+# get absolute parent directory path of current file
+CURPATH=`dirname $(realpath "$0")`
+cd $CURPATH
+
+# set environment with env config.
+set -a
+source ../env
+set +a
+
+# set pythonpath environment
+cd ..
+export PYTHONPATH=$PWD:$PWD/python-qa-tools
+echo $PYTHONPATH
+
+# check environment variables are set
+bash ./deps/env-check.sh $CURPATH
+
+python3 ./python-qa-tools/load-test/send_load.py $1 $2 $3 $4
