--- python-qa-tools/.env.example
@@ -1,9 +1,0 @@
-DAEMON=simd
-DENOM=stake
-CHAINID=test
-DAEMON_HOME=${HOME}/.simd
-GH_URL=https://github.com/cosmos/cosmos-sdk
-CHAIN_VERSION='v0.45.1'
-UPGRADE_NAME=test
-UPGRADE_VERSION='v0.46.0-alpha4'
-NUM_MSGS=30
