+++ python-qa-tools/core/keys.py
@@ -1,5 +1,7 @@
 import os, json
-from utils.commands import exec_command
+
+from utils import exec_command
+
 
 DAEMON = os.getenv('DAEMON')
 DAEMON_HOME = os.getenv('DAEMON_HOME')
