+++ python-qa-tools/modules/auth/query.py
@@ -1,10 +1,18 @@
-import os, json
-from utils.commands import exec_command
+import argparse, os, json
+
+from utils import exec_command
 
 DAEMON = os.getenv('DAEMON')
 RPC = os.getenv('RPC')
 
 # query account function will take the bech32 address as input and output the information of account.
+
+def account_type(address):
+    status, response = query_account(address)
+    if not status:
+        raise argparse.ArgumentTypeError(response)
+    return address
+
 def query_account(address):
     try:
         command = f"{DAEMON} query auth account {address} --node {RPC} --output json"
