--- python-qa-tools/requirements.txt
@@ -1,1 +1,0 @@
-python-dotenv==0.20.0
